Move topics subcommand to repos
To align with how the GitHub API is organised
