+++ src/commands/repos.js
@@ -2,9 +2,9 @@
  * This project follows the example provided in the Yargs documentation for command hierarchy and directory structure.
  * @see: https://github.com/yargs/yargs/blob/master/docs/advanced.md#commanddirdirectory-opts
  */
-exports.command = 'topics <subcommand> [...options]'
-exports.desc = 'Manage GitHub topics'
+exports.command = 'repos <subcommand> [...options]'
+exports.desc = 'Manage GitHub repositories'
 exports.builder = function(yargs) {
-	return yargs.commandDir('topics')
+	return yargs.commandDir('repos')
 }
 exports.handler = function() {}
