+++ src/commands/repos/remove-topics.js
@@ -1,12 +1,12 @@
 /**
- * @see: https://octokit.github.io/rest.js/#octokit-routes-repos-list-topics
- * List all topics of a repository.
- * const result = await octokit.repos.listTopics()
+ * @see: https://octokit.github.io/rest.js/#octokit-routes-repos-replace-topics
+ * Remove topics from a repository.
+ * const result = await octokit.repos.replaceTopics()
  */
 const flow = require('lodash.flow')
-const commonYargs = require('../../lib/common-yargs')
+const { withGitHubUrl, withTopics } = require('../../lib/common-yargs')
 const printOutput = require('../../lib/print-output')
-const { getTopics } = require('../../lib/topics')
+const { removeTopics } = require('../../lib/topics')
 
 /**
  * yargs builder function.
@@ -15,25 +15,28 @@ const { getTopics } = require('../../lib/topics')
  */
 const builder = yargs => {
 	const baseOptions = flow([
-		commonYargs.withGitHubUrl({
+		withGitHubUrl({
 			describe: 'The URL of the GitHub repository to list the topics of.',
+		}),
+		withTopics({
+			required: true,
 		}),
 	])
 	return baseOptions(yargs).example('github-url', 'Pattern: https://github.com/[owner]/[repository]')
 }
 
 /**
- * List all issues for a repository.
+ * Remove topics from a repository.
  *
  * @param {object} argv - argv parsed and filtered by yargs
  * @param {string} argv.token
  * @param {string} argv.json
  * @param {object} argv.githubUrl - The GitHub url parsed in the withGitHubUrl() yarg option into appropriate properties, such as `owner` and `repo`.
+ * @param {array} argv.topics - The topics to remove
  */
-const handler = async args => {
-	const { token, json, githubUrl } = args
+const handler = async ({ token, json, githubUrl, topics: topicsToRemove }) => {
 	try {
-		const topics = await getTopics({ githubUrl, token })
+		const topics = await removeTopics({ githubUrl, token, topics: topicsToRemove })
 		if (json) {
 			printOutput({ json, resource: topics })
 		} else {
@@ -46,8 +49,8 @@ const handler = async args => {
 }
 
 module.exports = {
-	command: 'list <github-url>',
-	desc: 'List all topics of a repository.',
+	command: 'remove-topics <github-url>',
+	desc: 'Remove topics from a repository.',
 	builder,
 	handler,
 }
